add cold load by press on PC12

+++ device/include/timestamp.h
@@ -7,9 +7,9 @@
 #include "cmsis_os2.h"
 
 static uint32_t getTimestampMs() {
-    // Convert ticks to milliseconds safely based on your RTOS clock rate
-    // (If configTICK_RATE_HZ is 1000, ms directly equals ticks)
-    return osKernelGetTickCount() * 1000 / osKernelGetTickFreq();
+    // 64-bit intermediate avoids overflowing the tick*1000 multiplication in
+    // 32 bits, which previously wrapped after ~71 minutes of uptime.
+    return (uint32_t)((uint64_t)osKernelGetTickCount() * 1000ULL / osKernelGetTickFreq());
 }
 
 static int64_t getTimestampNs() {
